Refactor consultee layout and page components
This commit refactors the consultee layout and page components. In the layout component, the searchParams prop has been removed as it is no longer used. In the page component, the params prop has been replaced with a Promise-based params prop to improve performance. Additionally, the redirect function has been updated to use the resolvedParams value instead of the original params value.